DELIA-65803 - Async retry logic to check netsrvmgr is active
Reason for change: Added asynchronous retry logic with thread to check the
availability of netsrvmgr in network manager plugin
Test Procedure: Check the status of network manager plugin is activated in
wpeframework
Risks: Low
Priority: P1

RDKTV-32272 - Wi-Fi Connection in Network page is showing 2.4Ghz when connected to 5Ghz network

Reason for change: The value 2 showed in the frequency is enum value of 5GHz. Changed the code to return the proper frequency value in MHz.
Test Procedure: Check with curl commands
Risks: Low
Priority: P1

+++ NetworkManager/NetworkManagerJsonRpc.cpp
@@ -869,7 +869,7 @@ namespace WPEFramework
                 response["bssid"] = ssidInfo.m_bssid;
                 response["securityMode"] = static_cast <int> (ssidInfo.m_securityMode);
                 response["signalStrength"] = ssidInfo.m_signalStrength;
-                response["frequency"] = static_cast <int> (ssidInfo.m_frequency);
+                response["frequency"] = ssidInfo.m_frequency;
                 response["rate"] = ssidInfo.m_rate;
                 response["noise"] = ssidInfo.m_noise;
                 response["success"] = true;
